Decouple DomainExtractor from Interpreter
In connector, we can always use an simpified extractor to assume engine
will always pass down pre-optimized predicate.
